refactor toolpadComponents
